Remove language field from Hook — hooks are opaque executables
abx-dl no longer cares about hook implementation language. Hooks are
black-box +x processes identified only by filename convention. Also
moves orchestrator logic into service classes and updates tests.
