+++ abx_dl/services/crawl_service.py
@@ -1,15 +1,15 @@
 """CrawlService — registers per-hook handlers for CrawlEvent and SnapshotEvent."""
 
 from pathlib import Path
-from typing import Any, ClassVar
+from typing import ClassVar
 
 from bubus import BaseEvent, EventBus
 
-from ..config import build_env_for_plugin
 from ..events import CrawlEvent, ProcessEvent, SnapshotEvent
 from ..models import Snapshot
 from ..plugins import Hook, Plugin
 from .base import BaseService
+from .machine_service import MachineService
 
 
 class CrawlService(BaseService):
@@ -25,14 +25,14 @@ class CrawlService(BaseService):
         url: str,
         snapshot: Snapshot,
         output_dir: Path,
-        shared_config: dict[str, Any],
+        machine: MachineService,
         crawl_hooks: list[tuple[Plugin, Hook]],
         snapshot_hooks: list[tuple[Plugin, Hook]],
     ):
         self.url = url
         self.snapshot = snapshot
         self.output_dir = output_dir
-        self.shared_config = shared_config
+        self.machine = machine
         self.crawl_hooks = crawl_hooks
         self.snapshot_hooks = snapshot_hooks
         # Don't call super().__init__ yet — we register dynamic handlers manually
@@ -56,21 +56,19 @@ class CrawlService(BaseService):
     def _make_hook_handler(self, plugin: Plugin, hook: Hook, event_cls: type[BaseEvent]):
         """Create an async handler that builds env and emits ProcessEvent."""
         async def handler(event: BaseEvent, _plugin=plugin, _hook=hook) -> None:
-            env = build_env_for_plugin(
-                _plugin.name, _plugin.config_schema, self.shared_config,
-                run_output_dir=self.output_dir,
-            )
+            env = self.machine.get_env_for_plugin(_plugin, run_output_dir=self.output_dir)
             timeout = int(env.get(f"{_plugin.name.upper()}_TIMEOUT", env.get('TIMEOUT', '60')))
             plugin_output_dir = self.output_dir / _plugin.name
             plugin_output_dir.mkdir(parents=True, exist_ok=True)
 
             # Queue-jump: ProcessEvent handler runs the subprocess synchronously
             await self.bus.emit(ProcessEvent(
-                url=self.url, snapshot_id=self.snapshot.id,
                 plugin_name=_plugin.name, hook_name=_hook.name,
-                hook_path=str(_hook.path), hook_language=_hook.language,
+                hook_path=str(_hook.path),
+                hook_args=[f'--url={self.url}', f'--snapshot-id={self.snapshot.id}'],
                 is_background=_hook.is_background,
-                output_dir=str(plugin_output_dir), env=env, timeout=timeout,
+                output_dir=str(plugin_output_dir), env=env,
+                snapshot_id=self.snapshot.id, timeout=timeout,
             ))
 
         return handler
