+++ tests/test_plugins.py
@@ -4,7 +4,7 @@ from abx_dl.plugins import discover_plugins, parse_hook_filename
 
 
 def test_parse_hook_filename_marks_bg_hooks() -> None:
-    assert parse_hook_filename('on_Snapshot__66_papersdl.finite.bg.py') == ('Snapshot', 6, 6, True, 'py')
+    assert parse_hook_filename('on_Snapshot__66_papersdl.finite.bg.py') == ('Snapshot', 6, 6, True)
 
 
 def test_discover_plugins_marks_papersdl_as_background() -> None:
@@ -17,6 +17,5 @@ def test_discover_plugins_marks_papersdl_as_background() -> None:
     )
 
     assert papersdl_hook.is_background is True
-    assert papersdl_hook.language == 'py'
     assert papersdl_hook.path.parent == plugins['papersdl'].path
     assert 'papersdl' in papersdl_hook.path.name
